Build parser once (not on every load); add test suite; fix rule ordering.
PEG grammars, unlike LALR grammars, have *ordered* choices.  This (for
example) means that a null production (which always succeeds) cannot be
the first choice in an ordered list, or else none of the other alternatives
will ever be tried.  Re-order some rules in the source grammar to accomodate
this.

Add a basic test suite for the parser.

Rename some files; the generated parser class is now `parser.js`.