Bug fixes and consistency improvements.
Ensure that we can parse what we emit; be more consistent in our use of
trailing spaces in the AST and the render stage (which also helps ensure
that we implement the identity function when rendering our output).

The translation of `\varcoppa` has been changed to `\mbox{\coppa}`; that
seems to be a bug in the ocamlvc.  The `\mathbb`, `\mathbf`, and `\mathrm`
functions were moved to the `fun_ar1nb` class (which means they don't need
surrounding braces when rendering).  The `\operatorname` function existed
in two different classes at once; move it to only `fun_ar1nb`.

Some bug fixes: LList wasn't handling constructing a singleton list correctly,
and fix the top-level FUN_INFIX/FUN_INFIXh and DECLh productions to construct
their lists correctly.  Move the FUN_AR1opt production higher to ensure it
gets tried before the FUN_AR1 production.  Add a production for parsing
`\mbox{\someoperator}`, which we can emit.

Added `options` field to `check`, to allow getting stack traces on parser
failure.