Move `contains_func` into the API.
Move the `contains_func` visitor from a test case example of the
low-level API into the API proper.  Extend it so that it can test
against a set of functions at once, instead of one at a time.

We will use this to implement the checks for the use of
specific TeX packages.