Add support for italic capital Greek letters
There is currently no way to write a bold italic capital Greek letter (e.g. for angular velocity vector).

The easiest way to achieve this in Math/MathJax is to use `\boldsymbol \varOmega` which is already supported by MathJax, but it is not recognized by texvc, so it is rejected.

This commit adds all symbols for capital Greek variables (`varDelta`, `\varOmega`, ...) supported by MathJax (see https://docs.mathjax.org/en/latest/tex.html#v) to the list of supported macros.