Implement array syntax object
Implement array syntax object as alternative to native js arrays.

Now the use of the new syntax element is optional. Refactoring to always use the new array element will come in a follow-up.

The new element is hidden in the tree visualization.
The root element in tree visualizations changes from root->array which is more appropriate since the "root" element could appear in non-root positions of the tree before.

Change-Id: I1c676e5352a4250891a5e1501b96231114b0f513